Make ::Job only visible to ::Manager
This will allow embedding a reference to the
manager's eventmgr object, which allows more
internal Job logic to be moved from ::Manager to ::Job.

A consequence of this change is the need to move
some standalone functions into ::Job, and to make
the manager unit tests part of a ::ManagerTest
friend class.